refactor(console): replace custom-user-fields angular js component with migrated
