+++ gravitee-apim-console-webui/src/management/settings/custom-user-fields/dialog/custom-user-fields-dialog.component.scss
