--- gravitee-apim-console-webui/src/management/settings/custom-user-fields/dialog/update.custom-user-field.dialog.controller.ts
@@ -1,42 +1,0 @@
-/*
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import CustomUserFieldsService from '../../../../services/custom-user-fields.service';
-
-function UpdateFieldDialogController(
-  CustomUserFieldsService: CustomUserFieldsService,
-  $mdDialog: angular.material.IDialogService,
-  field,
-  predefinedKeys,
-) {
-  this.updateAction = true;
-
-  this.field = CustomUserFieldsService.mapForUpdateDialog(field);
-  this.predefinedKeys = predefinedKeys;
-
-  this.cancel = () => {
-    $mdDialog.cancel();
-  };
-
-  this.save = () => {
-    CustomUserFieldsService.update(this.field).then((response) => {
-      $mdDialog.hide(response.data);
-    });
-  };
-}
-UpdateFieldDialogController.$inject = ['CustomUserFieldsService', '$mdDialog', 'field', 'predefinedKeys'];
-
-export default UpdateFieldDialogController;
